get last update

+++ Edit Components Application/Editor/UnitTests/EditorTests.cs
@@ -31,7 +31,6 @@ using System.Threading.Tasks;
 using System.Windows.Forms;
 using System.IO;
 using Editor;
-
 namespace UnitTests
 {
     [TestClass]
@@ -115,8 +114,60 @@ namespace UnitTests
             //Checks to see if the results from the file are the same as the data written in.
             Assert.AreEqual(sb.ToString(), results);
 
-            //Deletes the temporary file.
-            File.Delete(path);
+            //Make sure the test file wasn't deleted during the test
+            if (File.Exists(path))
+            {
+                //Deletes the temporary file.
+                File.Delete(path);
+            }
+
         }
+
+
+        /// <summary>
+        /// Checks to see if the desktop image can be replaced correctly
+        /// </summary>
+        [TestMethod]
+        public void CheckReplacingFile()
+        {
+            // Creates an instance of the editor
+            var editor = new uxEditor();
+
+            // The test image
+            string pathToTestImage = "../../application/src/Resources/test_background.png";
+
+            // The path to the desktop image
+            string pathToOriginal = "../../application/src/Resources/Title_Computer.png";
+
+            // The path to the backup image
+            string backup = "../../application/src/Resources/backup_of_desktop.png";
+            string testbackup = "../../application/src/Resources/test_backup.png";
+
+            // Checks to see if all the paths exist
+            if (File.Exists(pathToTestImage) && File.Exists(pathToOriginal) && File.Exists(backup))
+            {
+                // Attempts to change the desktop image (Note: I have made the original method to take in all 3 paths
+                // because testing only allows access to paths that have an extra ../
+                // To counteract this I created 3 new arguments  to the method
+                Assert.IsTrue(editor.UpdateDesktopImage(pathToTestImage,pathToOriginal,backup));
+
+                // If the change was success change the desktop image back to the original. 
+                editor.UpdateDesktopImage(backup, pathToOriginal, testbackup);
+
+                // Make sure the test file was not deleted during the test
+                if (File.Exists(testbackup))
+                {
+                    File.Delete(testbackup);
+                }
+               
+            }
+            else
+            {
+                // If the test files cannot be found fail the test
+                Assert.Fail();
+            }
+            
+        }
+
     }
 }
